imx-lib: Add 09.12.01 Release
This ebuild has the same caveats as the libz160-bin ebuild.  A lot of
work needs to be done to clean this up and make it acceptable.  .bin
files are getting installed into /usr/lib - they seem to be some sort of
firmware for the vpu.  It also includes things from
/usr/src/linux/include as well as /usr/src/linux/drivers/mxc...  This
needs to change.