xf86-video-imx: 2 patches for Xorg 1.8 compat.
Added 2 more patches.  One removes the Symbols and syms lines, and the
other kills off the Resources stuff behind some ifndef's (gathered via
this commit:
http://cgit.freedesktop.org/xorg/driver/xf86-video-intel/commit/?id=9a3b568d62a0b48f4a42ea5377740b2df1af432a

I've no idea if this is the proper way now, but it was the first URL
that I came across that was close to my issue.